Improve selection helpers
